Seperate keys.go into keypair and armour logic + move into crypto package
